Tank v0.3 & Minor Changes
- Fixed the stuttering movement of the tank thanks to Jay.
- Changed the health related variables from floats to integers as TF2 uses integers.
- Implemented Wander behaviour to the robots. Will use it on a sequence later.

+++ TF2 MvM AI Attempt/Assets/Scripts/Player/HealthBar.cs
@@ -5,12 +5,12 @@ public class HealthBar : MonoBehaviour
 {
     public Slider healthSlider;
 
-    public void SetSliderHP(float amount)
+    public void SetSliderHP(int amount)
     {
         healthSlider.value = amount;
     }
 
-    public void SetSliderMaxHP(float amount)
+    public void SetSliderMaxHP(int amount)
     {
         healthSlider.maxValue = amount;
         SetSliderHP(amount);
